+++ CHANGELOG.md
@@ -7,7 +7,10 @@ This file documents all notable changes to the GEOS-Chem Classic wrapper reposit
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased 14.1.0]
-## Changed
+### Added
+- Added SANITIZE CMake option
+
+### Changed
 - Updated GEOS-Chem submodule to 14.1.0
 - Updated HEMCO submodule to 3.6.0